+++ pysui/sui/sui_pgql/config/__init__.py
@@ -4,3 +4,6 @@
 # -*- coding: utf-8 -*-
 
 """Pysui Sui GraphQL Configuration Package."""
+
+from .pysui_config import PysuiConfiguration
+from .confmodel import PysuiConfigModel