+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - On PysuiConfiguration:
+  - `add_keys` adds a collection of keys and optional aliases
   - `new_profile` creates equivalant of Sui CLI new-env
   - `update_profile` updates profile values